+++ Pandora/Paradigm/Controlflow/Effect/Interpreted.hs
@@ -7,9 +7,10 @@ import Pandora.Pattern.Transformer.Liftable (Liftable (lift))
 type family Schematic (c :: (* -> *) -> k) (t :: * -> *) = (r :: (* -> *) -> * -> *) | r -> t
 
 class Interpreted t where
-	{-# MINIMAL run #-}
+	{-# MINIMAL run, unite #-}
 	type Primary t a :: *
 	run :: t a -> Primary t a
+	unite :: Primary t a -> t a
 
 via :: (Liftable t, Interpreted (t u), Interpreted (t v), Covariant u)
 	=> (t u a -> t v b) -> u a -> Primary (t v) b