+++ Pandora/Paradigm/Primary/Functor/Conclusion.hs
@@ -15,7 +15,7 @@ import Pandora.Pattern.Object.Chain (Chain ((<=>)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean (False))
 import Pandora.Paradigm.Primary.Object.Ordering (Ordering (Less, Greater))
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Schematic, Interpreted (Primary, run))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Schematic, Interpreted (Primary, run, unite))
 import Pandora.Paradigm.Controlflow.Effect.Transformer.Monadic (Monadic (wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
 import Pandora.Paradigm.Schemes.UT (UT (UT), type (<.:>))
@@ -78,6 +78,7 @@ fail _ (Success y) = Success y
 instance Interpreted (Conclusion e) where
 	type Primary (Conclusion e) a = Conclusion e a
 	run = identity
+	unite = identity
 
 type instance Schematic Monad (Conclusion e) = (<.:>) (Conclusion e)