+++ Pandora/Paradigm/Primary/Transformer/Kan.hs
@@ -3,7 +3,7 @@ module Pandora.Paradigm.Primary.Transformer.Kan where
 import Pandora.Pattern.Category ((.), ($))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite))
 import Pandora.Paradigm.Primary.Functor.Wye (Wye (Left, Right))
 
 data family Kan (v :: * -> k) (t :: * -> *) (u :: * -> *) b a
@@ -16,6 +16,7 @@ instance Contravariant (Kan Left t u b) where
 instance Interpreted (Kan Left t u b) where
 	type Primary (Kan Left t u b) a = (t b -> a) -> u b
 	run ~(Lan x) = x
+	unite = Lan
 
 data instance Kan Right t u b a = Ran ((a -> t b) -> u b)
 
@@ -25,3 +26,4 @@ instance Covariant (Kan Right t u b) where
 instance Interpreted (Kan Right t u b) where
 	type Primary (Kan Right t u b) a = (a -> t b) -> u b
 	run ~(Ran x) = x
+	unite = Ran