+++ Pandora/Paradigm/Schemes/UTU.hs
@@ -3,7 +3,7 @@ module Pandora.Paradigm.Schemes.UTU where
 import Pandora.Core.Functor (type (:.), type (:=))
 import Pandora.Pattern.Functor.Covariant (Covariant)
 import Pandora.Pattern.Functor.Contravariant (Contravariant)
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite))
 
 newtype UTU ct cu t u u' a = UTU (u :. t :. u' := a)
 
@@ -19,3 +19,4 @@ type (>.>:<.<) = UTU Contravariant Contravariant Contravariant
 instance Interpreted (UTU ct cu t u u') where
 	type Primary (UTU ct cu t u u') a = u :. t :. u' := a
 	run ~(UTU x) = x
+	unite = UTU