+++ Pandora/Paradigm/Schemes/U_T.hs
@@ -9,7 +9,7 @@ import Pandora.Pattern.Transformer.Liftable (Liftable (lift))
 import Pandora.Pattern.Transformer.Lowerable (Lowerable (lower))
 import Pandora.Pattern.Transformer.Hoistable (Hoistable (hoist))
 import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)), type (:*:))
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite))
 
 newtype U_T ct cu t u a = U_T (u a :*: t a)
 
@@ -21,6 +21,7 @@ type (>.:.<) = U_T Contravariant Contravariant
 instance Interpreted (U_T ct cu t u) where
 	type Primary (U_T ct cu t u) a = u a :*: t a
 	run ~(U_T x) = x
+	unite = U_T
 
 instance Avoidable t => Liftable (U_T Covariant Covariant t) where
 	lift :: Covariant u => u ~> t <.:.> u